!B (CE-10909) Fixed memory counter overflow when generating 1024 cubemap which gets 4x4 supersampling (Approved by nicolas) !B (CE-10907) Fixed crash when assert occurs during cubemap generation Review: ID 62991 by Author: Nicolas Schulz, Reviewer: Theodor Mader, Observer: Jan Hellwig, Observer: Antonios Ryakiotakis
Copied from Perforce
 Change: 1470230

+++ Code/CryEngine/RenderDll/Common/Textures/Texture.h
@@ -1823,7 +1823,7 @@ public:
 	static volatile size_t s_nStatsStreamPoolBoundMem;          // Amount of stream pool currently bound and in use by textures (avail + non avail)
 	static volatile size_t s_nStatsStreamPoolBoundPersMem;      // Amount of stream pool currently bound and in use by persistent texture mem (avail + non avail)
 	static volatile int    s_nStatsCurManagedNonStreamedTexMem;
-	static volatile int    s_nStatsCurDynamicTexMem;
+	static volatile size_t s_nStatsCurDynamicTexMem;
 	static volatile size_t s_nStatsStreamPoolWanted;
 	static bool            s_bStatsComputeStreamPoolWanted;
